@@ -1,39 +1,27 @@
 package com.authorization.server.security;
 
-import lombok.Cleanup;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.ApplicationRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.annotation.Order;
-import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.oauth2.core.AuthorizationGrantType;
-import org.springframework.security.oauth2.core.ClientAuthenticationMethod;
 import org.springframework.security.oauth2.core.OAuth2RefreshToken;
 import org.springframework.security.oauth2.core.OAuth2Token;
-import org.springframework.security.oauth2.core.oidc.OidcScopes;
 import org.springframework.security.oauth2.jwt.NimbusJwtEncoder;
-import org.springframework.security.oauth2.server.authorization.OAuth2Authorization;
 import org.springframework.security.oauth2.server.authorization.client.JdbcRegisteredClientRepository;
-import org.springframework.security.oauth2.server.authorization.client.RegisteredClient;
 import org.springframework.security.oauth2.server.authorization.client.RegisteredClientRepository;
 import org.springframework.security.oauth2.server.authorization.config.annotation.web.configurers.OAuth2AuthorizationServerConfigurer;
 import org.springframework.security.oauth2.server.authorization.settings.AuthorizationServerSettings;
-import org.springframework.security.oauth2.server.authorization.settings.ClientSettings;
-import org.springframework.security.oauth2.server.authorization.settings.TokenSettings;
 import org.springframework.security.oauth2.server.authorization.token.DelegatingOAuth2TokenGenerator;
 import org.springframework.security.oauth2.server.authorization.token.JwtEncodingContext;
 import org.springframework.security.oauth2.server.authorization.token.OAuth2TokenCustomizer;
 import org.springframework.security.oauth2.server.authorization.token.OAuth2TokenGenerator;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
 import org.springframework.security.web.authentication.SavedRequestAwareAuthenticationSuccessHandler;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
@@ -43,68 +31,43 @@ import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
 import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
-import java.io.IOException;
-import java.security.NoSuchAlgorithmException;
-import java.security.spec.InvalidKeySpecException;
-import java.time.Duration;
 import java.util.Arrays;
 import java.util.List;
-import java.util.UUID;
 import java.util.stream.Collectors;
 
-import static java.util.UUID.randomUUID;
-import static org.springframework.http.HttpMethod.POST;
+import static org.springframework.core.Ordered.HIGHEST_PRECEDENCE;
+import static org.springframework.http.HttpHeaders.*;
+import static org.springframework.http.HttpHeaders.ACCESS_CONTROL_ALLOW_CREDENTIALS;
+import static org.springframework.http.HttpMethod.*;
+import static org.springframework.http.HttpMethod.OPTIONS;
 import static org.springframework.security.oauth2.server.authorization.OAuth2TokenType.ACCESS_TOKEN;
-
-@Configuration
 @Slf4j
+@Configuration
 @RequiredArgsConstructor
 public class AuthorizationServerConfig {
+    // http://localhost:8080/.well-known/openid-configuration
+    private static final String AUTHORITY_KEY = "authorities";
     private final JwtConfiguration jwtConfiguration;
-    private static final String AUTHORITY_KEY= "authorities";
-
-    @Bean
-    public RegisteredClientRepository registeredClientRepository(JdbcTemplate jdbcTemplate) {
-        return new JdbcRegisteredClientRepository(jdbcTemplate);
-    }
-
-    @Bean
-    public AuthenticationSuccessHandler authenticationSuccessHandler(){
-        return new SavedRequestAwareAuthenticationSuccessHandler();
-    }
-    @Bean
-    public AuthorizationServerSettings authorizationServerSettings(){
-        return AuthorizationServerSettings.builder().build();
-    }
-
-    @Bean
-    public OAuth2TokenCustomizer<JwtEncodingContext> customizer(){
-        return context -> {
-            if (ACCESS_TOKEN.equals(context.getTokenType())){
-                context.getClaims().claims(claims->claims.put(AUTHORITY_KEY, getAuthorities(context)));
-            }
-        };
-    }
-
-
 
     @Bean
     @Order(1)
-    public SecurityFilterChain authorizationServerConfig(HttpSecurity http,
-                                                         RegisteredClientRepository registeredClientRepository) throws Exception {
+    public SecurityFilterChain oauth2ServerConfig(HttpSecurity http, RegisteredClientRepository registeredClientRepository, OAuth2TokenCustomizer<JwtEncodingContext> customizer) throws Exception {
         http.cors(corsConfigurer -> corsConfigurer.configurationSource(corsConfigurationSource()));
-        var authorizationConfig = OAuth2AuthorizationServerConfigurer.authorizationServer()
-                .tokenGenerator(tokenGenerator())
-                .clientAuthentication(authentication -> {
-                    authentication.authenticationConverter(new ClientRefreshTokenAuthenticationConverter());
-                    authentication.authenticationProvider(new ClientAuthenticationProvider(registeredClientRepository));
-                })
-                .oidc(Customizer.withDefaults());
-        http.securityMatcher(authorizationConfig.getEndpointsMatcher())
-                .with(authorizationConfig, Customizer.withDefaults())
-                .exceptionHandling(exceptions -> exceptions.accessDeniedPage("accessdenied")
-                        .defaultAuthenticationEntryPointFor(
-                                new LoginUrlAuthenticationEntryPoint("/login"), new MediaTypeRequestMatcher(MediaType.TEXT_HTML)));
+        var authorizationConfig = OAuth2AuthorizationServerConfigurer.authorizationServer();
+        http
+                .securityMatcher(authorizationConfig.getEndpointsMatcher())
+                .with(authorizationConfig, authorizationServer -> authorizationServer
+                                .oidc(Customizer.withDefaults())
+                                .authorizationServerSettings(authorizationServerSettings())
+                                .registeredClientRepository(registeredClientRepository)
+                                .tokenGenerator(tokenGenerator())
+                                .clientAuthentication(authentication -> {
+                                    authentication.authenticationConverter(new ClientRefreshTokenAuthenticationConverter());
+                                    authentication.authenticationProvider(new ClientAuthenticationProvider(registeredClientRepository));
+                        }))
+                .authorizeHttpRequests(authorize -> authorize.anyRequest().authenticated());
+        http.exceptionHandling(exceptions -> exceptions.accessDeniedPage("/accessdenied")
+                        .defaultAuthenticationEntryPointFor(new LoginUrlAuthenticationEntryPoint("/login"), new MediaTypeRequestMatcher(MediaType.TEXT_HTML)));
         return http.build();
     }
 
@@ -112,45 +75,67 @@ public class AuthorizationServerConfig {
     @Order(2)
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         http.cors(corsConfigurer -> corsConfigurer.configurationSource(corsConfigurationSource()));
-        http.authorizeHttpRequests(authorize->authorize
-                .requestMatchers("/login").permitAll()
-                .requestMatchers(POST, "logout").permitAll()
+        http.authorizeHttpRequests(authorize -> authorize
+                .requestMatchers("/login", "/oauth2/authorize/**", "oauth2/authorize").permitAll()
+                .requestMatchers(POST, "/logout").permitAll()
                 .requestMatchers("/mfa").hasAuthority("MFA_REQUIRED")
                 .anyRequest().authenticated());
-        http.formLogin(login->login
+        http.formLogin(login -> login
                 .loginPage("/login")
                 .successHandler(new MfaAuthenticationHandler("/mfa", "MFA_REQUIRED"))
-                .failureHandler(new SimpleUrlAuthenticationFailureHandler("/login?error")));
-        http.logout(logout->logout
-                .logoutSuccessUrl("/localhost:3000")
+                .failureHandler(new SimpleUrlAuthenticationFailureHandler("/login")));
+        http.logout(logout -> logout.logoutSuccessUrl("http://localhost:3000")
                 .addLogoutHandler(new CookieClearingLogoutHandler("JSESSIONID")));
         return http.build();
     }
+
     @Bean
-    public OAuth2TokenGenerator<? extends OAuth2Token> tokenGenerator() throws NoSuchAlgorithmException, IOException, InvalidKeySpecException {
-        var jwtGenerator = UserJwtGenerator.init(new NimbusJwtEncoder(jwtConfiguration.jwkSource()));
+    public OAuth2TokenGenerator<? extends OAuth2Token> tokenGenerator() {
+        var jwtGenerator = UserJwtGenerator.init(new NimbusJwtEncoder((jwtConfiguration.jwkSource())));
         jwtGenerator.setJwtCustomizer(customizer());
-        OAuth2TokenGenerator<OAuth2RefreshToken> oAuth2RefreshTokenOAuth2TokenGenerator = new ClientOAuth2TokenGenerator();
-        return new DelegatingOAuth2TokenGenerator(jwtGenerator,oAuth2RefreshTokenOAuth2TokenGenerator);
+        OAuth2TokenGenerator<OAuth2RefreshToken> refreshTokenOAuth2TokenGenerator = new ClientOAuth2RefreshTokenGenerator();
+        return new DelegatingOAuth2TokenGenerator(jwtGenerator, refreshTokenOAuth2TokenGenerator);
+    }
+
+    @Bean
+    public SavedRequestAwareAuthenticationSuccessHandler authenticationSuccessHandler() {
+        return new SavedRequestAwareAuthenticationSuccessHandler();
+    }
+
+    @Bean
+    public AuthorizationServerSettings authorizationServerSettings() {
+        return AuthorizationServerSettings.builder().build();
+    }
+
+    @Bean
+    public OAuth2TokenCustomizer<JwtEncodingContext> customizer() {
+        return context -> {
+            if(ACCESS_TOKEN.equals(context.getTokenType())) {
+                context.getClaims().claims(claims -> claims.put(AUTHORITY_KEY, getAuthorities(context)));
+            }
+        };
+    }
+
+    @Bean
+    public RegisteredClientRepository registeredClientRepository(JdbcTemplate jdbcTemplate) {
+        return new JdbcRegisteredClientRepository(jdbcTemplate);
     }
 
     @Bean
     public CorsConfigurationSource corsConfigurationSource() {
         var corsConfiguration = new CorsConfiguration();
         corsConfiguration.setAllowCredentials(true);
-        corsConfiguration.setAllowedOrigins(List.of("*"));
-        corsConfiguration.setAllowedHeaders(Arrays.asList("*"));
-        corsConfiguration.setExposedHeaders(Arrays.asList("*"));
-        corsConfiguration.setAllowedMethods(Arrays.asList("GET", "POST", "PUT", "DELETE", "OPTIONS"));
+        corsConfiguration.setAllowedOrigins(List.of("http://192.168.1.157:3000", "http://localhost:3000", "http://192.168.1.159:3000", "http://localhost:4200", "100.14.214.212:3000",  "http://localhost:4200", "http://localhost:3000", "http://192.168.1.216:3000"));
+        corsConfiguration.setAllowedHeaders(Arrays.asList(ORIGIN, ACCESS_CONTROL_ALLOW_ORIGIN, CONTENT_TYPE, ACCEPT, AUTHORIZATION, "X_REQUESTED_WITH", ACCESS_CONTROL_REQUEST_METHOD, ACCESS_CONTROL_REQUEST_HEADERS, ACCESS_CONTROL_ALLOW_CREDENTIALS));
+        corsConfiguration.setExposedHeaders(Arrays.asList(ORIGIN, ACCESS_CONTROL_ALLOW_ORIGIN, CONTENT_TYPE, ACCEPT, AUTHORIZATION, "X_REQUESTED_WITH", ACCESS_CONTROL_REQUEST_METHOD, ACCESS_CONTROL_REQUEST_HEADERS, ACCESS_CONTROL_ALLOW_CREDENTIALS));
+        corsConfiguration.setAllowedMethods(Arrays.asList(GET.name(), POST.name(), PUT.name(), PATCH.name(), DELETE.name(), OPTIONS.name()));
         corsConfiguration.setMaxAge(3600L);
         var source = new UrlBasedCorsConfigurationSource();
         source.registerCorsConfiguration("/**", corsConfiguration);
         return source;
     }
+
     private String getAuthorities(JwtEncodingContext context) {
-        return context
-                .getPrincipal()
-                .getAuthorities()
-                .stream().map(GrantedAuthority::getAuthority).collect(Collectors.joining(","));
+        return context.getPrincipal().getAuthorities().stream().map(GrantedAuthority::getAuthority).collect(Collectors.joining(","));
     }
 }
