@@ -1,15 +1,13 @@
 package com.authorization.server.controller;
-
-import com.authorization.server.entity.MfaAuthentication;
-import com.authorization.server.entity.User;
 import com.authorization.server.exception.ApiException;
+import com.authorization.server.model.User;
+import com.authorization.server.security.MfaAuthentication;
 import com.authorization.server.service.UserService;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.AllArgsConstructor;
-import org.apache.hc.core5.http.io.HttpServerRequestHandler;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.CurrentSecurityContext;
@@ -18,7 +16,6 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.WebAttributes;
 import org.springframework.security.web.authentication.AuthenticationFailureHandler;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.security.web.authentication.SavedRequestAwareAuthenticationSuccessHandler;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationFailureHandler;
 import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
 import org.springframework.security.web.context.SecurityContextRepository;
@@ -47,21 +44,6 @@ public class LoginController {
         return "login";
     }
 
-    @GetMapping("/logout")
-    public String logout() {
-        return "logout";
-    }
-    @RequestMapping("/error")
-    public String error(HttpServletRequest request, HttpServletResponse response, Model model, Exception exception) {
-        var errorException = (Exception) request.getAttribute(RequestDispatcher.ERROR_EXCEPTION);
-        if (errorException instanceof ApiException || errorException instanceof BadCredentialsException){
-            request.getSession().setAttribute(WebAttributes.AUTHENTICATION_EXCEPTION, errorException);
-            return "login";
-        }
-        model.addAttribute("message", getMessage(request));
-        return "error";
-    }
-
     @GetMapping("/mfa")
     public String mfa(Model model, @CurrentSecurityContext SecurityContext context) {
         model.addAttribute("email", getAuthenticatedUser(context.getAuthentication()));
@@ -71,28 +53,39 @@ public class LoginController {
     @PostMapping("/mfa")
     public void validateCode(@RequestParam("code") String code, HttpServletRequest request, HttpServletResponse response, @CurrentSecurityContext SecurityContext context) throws ServletException, IOException {
         var user = getUser(context.getAuthentication());
-        if (userService.verifyQrCode(user.getUserUuid(), code)){
-            this.authenticationSuccessHandler.onAuthenticationSuccess(request,response,getAuthentication(request, response));
+        if(userService.verifyQrCode(user.getUserUuid(), code)) {
+            this.authenticationSuccessHandler.onAuthenticationSuccess(request, response, getAuthentication(request, response));
             return;
         }
-        this.authenticationFailureHandler.onAuthenticationFailure(request, response, new BadCredentialsException("Invalid QR Code. Please try again"));
+        this.authenticationFailureHandler.onAuthenticationFailure(request, response, new BadCredentialsException("Invalid QR code. Please try again."));
+    }
+
+    @GetMapping("/logout")
+    public String logout() {
+        return "logout";
+    }
+
+    @GetMapping("/error")
+    public String error(HttpServletRequest request, HttpServletResponse response, Model model, Exception exception) {
+        var errorException = (Exception) request.getAttribute(RequestDispatcher.ERROR_EXCEPTION);
+        if(errorException instanceof ApiException || errorException instanceof BadCredentialsException) {
+            request.getSession().setAttribute(WebAttributes.AUTHENTICATION_EXCEPTION, errorException);
+            return "login";
+        }
+        model.addAttribute("message", getMessage(request));
+        return "error";
     }
 
     private Authentication getAuthentication(HttpServletRequest request, HttpServletResponse response) {
-            SecurityContext securityContext = SecurityContextHolder.getContext();
-            MfaAuthentication mfaAuthentication = (MfaAuthentication)securityContext.getAuthentication();
-            securityContext.setAuthentication(mfaAuthentication);
-            SecurityContextHolder.setContext(securityContext);
-            securityContextRepository.saveContext(securityContext, request,response);
-            return mfaAuthentication.getPrimaryAuthentication();
-        }
+        SecurityContext securityContext = SecurityContextHolder.getContext();
+        MfaAuthentication mfaAuthentication = (MfaAuthentication) securityContext.getAuthentication();
+        securityContext.setAuthentication(mfaAuthentication);
+        SecurityContextHolder.setContext(securityContext);
+        securityContextRepository.saveContext(securityContext, request, response);
+        return mfaAuthentication.getPrimaryAuthentication();
+    }
     private Object getAuthenticatedUser(Authentication authentication) {
         return ((User) authentication.getPrincipal()).getEmail();
     }
 
 }
-
-
-
-
-
