@@ -1,5 +1,5 @@
 package com.authorization.server.exception;
 
-public class ApiException extends RuntimeException{
-    public  ApiException(String message){super(message);}
+public class ApiException extends RuntimeException {
+    public ApiException(String message) { super(message); }
 }
