@@ -1,7 +1,7 @@
 package com.authorization.server.repository.implementation;
 
-import com.authorization.server.entity.User;
 import com.authorization.server.exception.ApiException;
+import com.authorization.server.model.User;
 import com.authorization.server.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -12,10 +12,9 @@ import org.springframework.stereotype.Service;
 import java.util.Map;
 
 import static com.authorization.server.query.UserQuery.*;
-
+@Slf4j
 @Service
 @RequiredArgsConstructor
-@Slf4j
 public class UserRepositoryImpl implements UserRepository {
     private final JdbcClient jdbc;
 
@@ -25,10 +24,10 @@ public class UserRepositoryImpl implements UserRepository {
             return jdbc.sql(SELECT_USER_BY_USER_UUID_QUERY).param("userUuid", userUuid).query(User.class).single();
         } catch (EmptyResultDataAccessException exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("No user found by UUID %s ", userUuid ));
+            throw new ApiException(String.format("No user found by UUID %s", userUuid));
         } catch (Exception exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("An error occurred. Please try again "));
+            throw new ApiException("An error occurred. Please try again.");
         }
     }
 
@@ -38,10 +37,10 @@ public class UserRepositoryImpl implements UserRepository {
             return jdbc.sql(SELECT_USER_BY_EMAIL_QUERY).param("email", email).query(User.class).single();
         } catch (EmptyResultDataAccessException exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("No user found by email %s ", email ));
+            throw new ApiException(String.format("No user found by email %s", email));
         } catch (Exception exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("An error occurred. Please try again "));
+            throw new ApiException("An error occurred. Please try again.");
         }
     }
 
@@ -51,10 +50,10 @@ public class UserRepositoryImpl implements UserRepository {
             jdbc.sql(RESET_LOGIN_ATTEMPTS_QUERY).param("userUuid", userUuid).update();
         } catch (EmptyResultDataAccessException exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("No user found by userUuid %s ", userUuid ));
+            throw new ApiException(String.format("No user found by UUID %s", userUuid));
         } catch (Exception exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("An error occurred. Please try again "));
+            throw new ApiException("An error occurred. Please try again.");
         }
     }
 
@@ -64,12 +63,11 @@ public class UserRepositoryImpl implements UserRepository {
             jdbc.sql(UPDATE_LOGIN_ATTEMPTS_QUERY).param("email", email).update();
         } catch (EmptyResultDataAccessException exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("No user found by email %s ", email ));
+            throw new ApiException(String.format("No user found by email %s", email));
         } catch (Exception exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("An error occurred. Please try again "));
+            throw new ApiException("An error occurred. Please try again.");
         }
-
     }
 
     @Override
@@ -78,23 +76,23 @@ public class UserRepositoryImpl implements UserRepository {
             jdbc.sql(SET_LAST_LOGIN_QUERY).param("userId", userId).update();
         } catch (EmptyResultDataAccessException exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("No user found by user Id %s ", userId ));
+            throw new ApiException(String.format("No user found user ID %s", userId));
         } catch (Exception exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("An error occurred. Please try again "));
+            throw new ApiException("An error occurred. Please try again.");
         }
     }
 
     @Override
     public void addLoginDevice(Long userId, String device, String client, String ipAddress) {
         try {
-            jdbc.sql(INSERT_NEW_DEVICE_QUERY).params(Map.of("userId", userId, "device",device, "client", client,"ipAddress",ipAddress)).update();
+            jdbc.sql(INSERT_NEW_DEVICE_QUERY).params(Map.of("userId", userId, "device", device, "client", client, "ipAddress", ipAddress)).update();
         } catch (EmptyResultDataAccessException exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("No user found by user Id %s ", userId ));
+            throw new ApiException(String.format("No user found user ID %s", userId));
         } catch (Exception exception) {
             log.error(exception.getMessage());
-            throw new ApiException(String.format("An error occurred. Please try again "));
+            throw new ApiException("An error occurred. Please try again.");
         }
     }
 }
