@@ -1,5 +1,4 @@
 package com.authorization.server.security;
-
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.oauth2.core.AuthorizationGrantType;
@@ -7,9 +6,9 @@ import org.springframework.security.oauth2.core.endpoint.OAuth2ParameterNames;
 import org.springframework.security.web.authentication.AuthenticationConverter;
 import org.springframework.stereotype.Component;
 import org.springframework.util.StringUtils;
-
 @Component
 public class ClientRefreshTokenAuthenticationConverter implements AuthenticationConverter {
+
     @Override
     public Authentication convert(HttpServletRequest request) {
         var grantType = request.getParameter(OAuth2ParameterNames.GRANT_TYPE);
@@ -17,7 +16,7 @@ public class ClientRefreshTokenAuthenticationConverter implements Authentication
             return null;
         }
         var clientId = request.getParameter(OAuth2ParameterNames.CLIENT_ID);
-        if (StringUtils.hasText(clientId)) {
+        if (!StringUtils.hasText(clientId)) {
             return null;
         }
         return new ClientRefreshTokenAuthentication(clientId);
