@@ -1,6 +1,5 @@
 package com.authorization.server.security;
 
-import com.authorization.server.utils.UserUtils;
 import org.springframework.lang.Nullable;
 import org.springframework.security.core.session.SessionInformation;
 import org.springframework.security.oauth2.core.AuthorizationGrantType;
@@ -10,7 +9,11 @@ import org.springframework.security.oauth2.core.oidc.OidcIdToken;
 import org.springframework.security.oauth2.core.oidc.endpoint.OidcParameterNames;
 import org.springframework.security.oauth2.jose.jws.JwsAlgorithm;
 import org.springframework.security.oauth2.jose.jws.SignatureAlgorithm;
-import org.springframework.security.oauth2.jwt.*;
+import org.springframework.security.oauth2.jwt.JwsHeader;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.security.oauth2.jwt.JwtClaimsSet;
+import org.springframework.security.oauth2.jwt.JwtEncoder;
+import org.springframework.security.oauth2.jwt.JwtEncoderParameters;
 import org.springframework.security.oauth2.server.authorization.OAuth2TokenType;
 import org.springframework.security.oauth2.server.authorization.client.RegisteredClient;
 import org.springframework.security.oauth2.server.authorization.settings.OAuth2TokenFormat;
@@ -21,12 +24,11 @@ import org.springframework.security.oauth2.server.authorization.token.OAuth2Toke
 import org.springframework.util.Assert;
 import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
-
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
 import java.util.Collections;
 import java.util.UUID;
-
+import static com.authorization.server.utils.UserUtils.getUser;
 public class UserJwtGenerator implements OAuth2TokenGenerator<Jwt> {
     private final JwtEncoder jwtEncoder;
     private OAuth2TokenCustomizer<JwtEncodingContext> jwtCustomizer;
@@ -35,7 +37,6 @@ public class UserJwtGenerator implements OAuth2TokenGenerator<Jwt> {
         Assert.notNull(jwtEncoder, "jwtEncoder cannot be null");
         this.jwtEncoder = jwtEncoder;
     }
-
 
     @Nullable
     @Override
@@ -64,7 +65,7 @@ public class UserJwtGenerator implements OAuth2TokenGenerator<Jwt> {
                 if (StringUtils.hasText(issuer)) {
                     claimsBuilder.issuer(issuer);
                 }
-                claimsBuilder.subject(UserUtils.getUser(context.getPrincipal()).getUserUuid()).audience(Collections.singletonList(registeredClient.getClientId())).issuedAt(issuedAt).expiresAt(expiresAt).id(UUID.randomUUID().toString());
+                claimsBuilder.subject(getUser(context.getPrincipal()).getUserUuid()).audience(Collections.singletonList(registeredClient.getClientId())).issuedAt(issuedAt).expiresAt(expiresAt).id(UUID.randomUUID().toString());
                 SessionInformation sessionInformation;
                 if (OAuth2TokenType.ACCESS_TOKEN.equals(context.getTokenType())) {
                     claimsBuilder.notBefore(issuedAt);
@@ -74,22 +75,21 @@ public class UserJwtGenerator implements OAuth2TokenGenerator<Jwt> {
                 } else if (OidcParameterNames.ID_TOKEN.equals(context.getTokenType().getValue())) {
                     claimsBuilder.claim(IdTokenClaimNames.AZP, registeredClient.getClientId());
                     if (AuthorizationGrantType.AUTHORIZATION_CODE.equals(context.getAuthorizationGrantType())) {
-                        OAuth2AuthorizationRequest authorizationRequest = (OAuth2AuthorizationRequest) context.getAuthorization().getAttribute(OAuth2AuthorizationRequest.class.getName());
-                        String nonce = (String) authorizationRequest.getAdditionalParameters().get("nonce");
+                        OAuth2AuthorizationRequest authorizationRequest = (OAuth2AuthorizationRequest)context.getAuthorization().getAttribute(OAuth2AuthorizationRequest.class.getName());
+                        String nonce = (String)authorizationRequest.getAdditionalParameters().get("nonce");
                         if (StringUtils.hasText(nonce)) {
                             claimsBuilder.claim(IdTokenClaimNames.NONCE, nonce);
                         }
-                        sessionInformation = (SessionInformation) context.get(SessionInformation.class);
+                        sessionInformation = (SessionInformation)context.get(SessionInformation.class);
                         if (sessionInformation != null) {
                             claimsBuilder.claim("sid", sessionInformation.getSessionId());
                             claimsBuilder.claim(IdTokenClaimNames.AUTH_TIME, sessionInformation.getLastRequest());
                         }
                     } else if (AuthorizationGrantType.REFRESH_TOKEN.equals(context.getAuthorizationGrantType())) {
-                        OidcIdToken currentIdToken = (OidcIdToken) context.getAuthorization().getToken(OidcIdToken.class).getToken();
+                        OidcIdToken currentIdToken = (OidcIdToken)context.getAuthorization().getToken(OidcIdToken.class).getToken();
                         if (currentIdToken.hasClaim("sid")) {
                             claimsBuilder.claim("sid", currentIdToken.getClaim("sid"));
                         }
-
                         if (currentIdToken.hasClaim(IdTokenClaimNames.AUTH_TIME)) {
                             claimsBuilder.claim(IdTokenClaimNames.AUTH_TIME, currentIdToken.getClaim(IdTokenClaimNames.AUTH_TIME));
                         }
@@ -98,8 +98,9 @@ public class UserJwtGenerator implements OAuth2TokenGenerator<Jwt> {
 
                 JwsHeader.Builder jwsHeaderBuilder = JwsHeader.with(jwsAlgorithm);
                 if (this.jwtCustomizer != null) {
-                    JwtEncodingContext.Builder jwtContextBuilder = JwtEncodingContext.with(jwsHeaderBuilder, claimsBuilder).registeredClient(context
-                                    .getRegisteredClient()).principal(context.getPrincipal())
+                    JwtEncodingContext.Builder jwtContextBuilder = JwtEncodingContext.with(jwsHeaderBuilder, claimsBuilder)
+                            .registeredClient(context.getRegisteredClient())
+                            .principal(context.getPrincipal())
                             .authorizationServerContext(context.getAuthorizationServerContext())
                             .authorizedScopes(context.getAuthorizedScopes())
                             .tokenType(context.getTokenType())
@@ -107,22 +108,18 @@ public class UserJwtGenerator implements OAuth2TokenGenerator<Jwt> {
                     if (context.getAuthorization() != null) {
                         jwtContextBuilder.authorization(context.getAuthorization());
                     }
-
                     if (context.getAuthorizationGrant() != null) {
                         jwtContextBuilder.authorizationGrant(context.getAuthorizationGrant());
                     }
-
                     if (OidcParameterNames.ID_TOKEN.equals(context.getTokenType().getValue())) {
-                        sessionInformation = (SessionInformation) context.get(SessionInformation.class);
+                        sessionInformation = (SessionInformation)context.get(SessionInformation.class);
                         if (sessionInformation != null) {
                             jwtContextBuilder.put(SessionInformation.class, sessionInformation);
                         }
                     }
-
                     JwtEncodingContext jwtContext = jwtContextBuilder.build();
                     this.jwtCustomizer.customize(jwtContext);
                 }
-
                 JwsHeader jwsHeader = jwsHeaderBuilder.build();
                 JwtClaimsSet claims = claimsBuilder.build();
                 Jwt jwt = this.jwtEncoder.encode(JwtEncoderParameters.from(jwsHeader, claims));
