@@ -1,6 +1,6 @@
 package com.authorization.server.service;
 
-import com.authorization.server.entity.User;
+import com.authorization.server.model.User;
 
 public interface UserService {
     User getUserByEmail(String email);
@@ -8,5 +8,5 @@ public interface UserService {
     void updateLoginAttempts(String email);
     void setLastLogin(Long userId);
     void addLoginDevice(Long userId, String deviceName, String client, String ipAddress);
-    boolean verifyQrCode(String userUuid,String code);
+    boolean verifyQrCode(String userUuid, String code);
 }
