@@ -1,6 +1,5 @@
 package com.authorization.server.utils;
-
-import com.authorization.server.entity.User;
+import com.authorization.server.model.User;
 import dev.samstevens.totp.code.CodeGenerator;
 import dev.samstevens.totp.code.CodeVerifier;
 import dev.samstevens.totp.code.DefaultCodeGenerator;
@@ -9,19 +8,20 @@ import dev.samstevens.totp.time.SystemTimeProvider;
 import dev.samstevens.totp.time.TimeProvider;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.oauth2.server.authorization.authentication.OAuth2AuthorizationCodeAuthenticationToken;
+import org.springframework.security.oauth2.server.authorization.authentication.OAuth2AuthorizationCodeRequestAuthenticationToken;
 
 public class UserUtils {
-public static boolean verifyQrCode(String secret, String code){
-    TimeProvider timeProvider = new SystemTimeProvider();
-    CodeGenerator codeGenerator = new DefaultCodeGenerator();
-    CodeVerifier verifier = new DefaultCodeVerifier(codeGenerator,timeProvider);
-    return verifier.isValidCode(secret,code);
-}
-    public static User getUser(Authentication authentication){
-        if (authentication instanceof OAuth2AuthorizationCodeAuthenticationToken){
-            var usernamePasswordAuthenticationToken = (UsernamePasswordAuthenticationToken)authentication.getPrincipal();
-            return (User) usernamePasswordAuthenticationToken.getPrincipal();
+    public static boolean verifyQrCode(String secret, String code) {
+        TimeProvider timeProvider = new SystemTimeProvider();
+        CodeGenerator codeGenerator = new DefaultCodeGenerator();
+        CodeVerifier verifier = new DefaultCodeVerifier(codeGenerator, timeProvider);
+        return verifier.isValidCode(secret, code);
+    }
+
+    public static User getUser(Authentication authentication) {
+        if(authentication instanceof OAuth2AuthorizationCodeRequestAuthenticationToken) {
+            var usernamePasswordAuthenticationToken = (UsernamePasswordAuthenticationToken) authentication.getPrincipal();
+                return (User) usernamePasswordAuthenticationToken.getPrincipal();
         }
         return (User) authentication.getPrincipal();
     }
