@@ -1,11 +1,11 @@
 package com.discovery.constants;
 
 public class Roles {
+    //Roles
     public static final String USER = "USER";
     public static final String TECH_SUPPORT = "TECH_SUPPORT";
     public static final String MANAGER = "MANAGER";
     public static final String ADMIN = "ADMIN";
     public static final String SUPER_ADMIN = "SUPER_ADMIN";
     public static final String APP_READ = "app:read";
-
 }
