@@ -8,13 +8,21 @@ import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.stereotype.Component;
 
 import java.io.IOException;
+
 @Component
 public class DiscoveryAccessDeniedHandler implements AccessDeniedHandler {
+
     @Override
     public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException exception) throws IOException, ServletException {
         response.setStatus(HttpServletResponse.SC_FORBIDDEN);
         var writer = response.getWriter();
-        writer.println("Http status 403 -" + exception.getMessage() );
-
+        writer.println("HTTP Status 403 - " + exception.getMessage());
+        //response.setContentType("TEXT/HTML");
+        /*writer.println("""
+                <html>
+                <head>
+                </head>
+                </html>
+                """);*/
     }
 }
