@@ -1,5 +1,6 @@
 package com.discovery.security;
 
+
 import com.discovery.repository.UserRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.security.core.userdetails.User;
@@ -10,21 +11,16 @@ import org.springframework.stereotype.Service;
 
 import static org.springframework.security.core.authority.AuthorityUtils.commaSeparatedStringToAuthorityList;
 
+
+
 @Service
 @AllArgsConstructor
 public class DiscoveryUserDetailsService implements UserDetailsService {
-
     private final UserRepository userRepository;
+
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
         var user = userRepository.getUserByUsername(username);
-        return new User(
-                user.getUsername(),
-                user.getPassword(),
-                user.isEnabled(),
-                user.isAccountNonExpired(),
-                user.isCredentialsNonExpired(),
-                user.isAccountNonLocked(),
-                commaSeparatedStringToAuthorityList(user.getRole() + "," +user.getAuthorities()));
+        return new User(user.getUsername(), user.getPassword(), user.isEnabled(), user.isAccountNonExpired(), user.isCredentialsNonExpired(), user.isAccountNonLocked(), commaSeparatedStringToAuthorityList(user.getRole() + "," + user.getAuthorities()));
     }
 }
