@@ -1,12 +1,13 @@
 package com.gateway.security;
 
-
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
+import org.springframework.stereotype.Component;
 
 import static org.springframework.security.core.authority.AuthorityUtils.commaSeparatedStringToAuthorityList;
 
+@Component
 public class JwtConverter implements Converter<Jwt, JwtAuthenticationToken> {
     private static final String AUTHORITY_KEY = "authorities";
 
@@ -16,5 +17,4 @@ public class JwtConverter implements Converter<Jwt, JwtAuthenticationToken> {
         var authorities = commaSeparatedStringToAuthorityList(claims);
         return new JwtAuthenticationToken(jwt, authorities, jwt.getSubject());
     }
-
 }
