@@ -1,5 +1,4 @@
 package com.gateway.utils;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.exc.MismatchedInputException;
 import com.gateway.domain.Response;
@@ -8,30 +7,32 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.authentication.*;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.CredentialsExpiredException;
+import org.springframework.security.authentication.DisabledException;
+import org.springframework.security.authentication.InsufficientAuthenticationException;
+import org.springframework.security.authentication.LockedException;
 import org.springframework.security.oauth2.server.resource.InvalidBearerTokenException;
-
-import java.util.Collections;
 import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
-
 import static java.time.LocalTime.now;
-import static org.apache.commons.lang.exception.ExceptionUtils.getRootCauseMessage;
+import static java.util.Collections.*;
+import static org.apache.commons.lang3.exception.ExceptionUtils.getRootCauseMessage;
 import static org.springframework.http.HttpStatus.*;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 
-public class RequestUtils {
-    private static final BiConsumer<HttpServletResponse, Response> writeResponse =
-            (servletResponse, response) -> {
-                try {
-                    var outputStream = servletResponse.getOutputStream();
-                    new ObjectMapper().writeValue(outputStream, response);
-                    outputStream.flush();
 
-                } catch (Exception exception) {
-                    throw new ApiException(exception.getMessage());
-                }
-            };
+public class RequestUtils {
+
+    private static final BiConsumer<HttpServletResponse, Response> writeResponse = (servletResponse, response) -> {
+        try {
+            var outputStream = servletResponse.getOutputStream();
+            new ObjectMapper().writeValue(outputStream, response);
+            outputStream.flush();
+        } catch (Exception exception) {
+            throw new ApiException(exception.getMessage());
+        }
+    };
 
     public static void handleErrorResponse(HttpServletRequest request, HttpServletResponse response, Exception exception) {
         if (exception instanceof AccessDeniedException) {
@@ -49,32 +50,32 @@ public class RequestUtils {
         } else if (exception instanceof DisabledException || exception instanceof LockedException || exception instanceof BadCredentialsException || exception instanceof CredentialsExpiredException || exception instanceof ApiException) {
             var apiResponse = getErrorResponse(request, response, exception, BAD_REQUEST);
             writeResponse.accept(response, apiResponse);
-        }else {
+        } else {
             var apiResponse = getErrorResponse(request, response, exception, INTERNAL_SERVER_ERROR);
             writeResponse.accept(response, apiResponse);
         }
     }
 
     private static final BiFunction<Exception, HttpStatus, String> errorReason = (exception, httpStatus) -> {
-        if (httpStatus.isSameCodeAs(FORBIDDEN)) {
-            return "You do not have enough permission";
+        if(httpStatus.isSameCodeAs(FORBIDDEN)) {
+            return "You don't have enough permission";
         }
-        if (httpStatus.isSameCodeAs(UNAUTHORIZED)) {
-            return exception.getMessage().contains("Jwt expired at") ? "You session has expired" : "You are not logged in";
+        if(httpStatus.isSameCodeAs(UNAUTHORIZED)) {
+            return exception.getMessage().contains("Jwt expired at") ? "Your session has expired" : "You are not logged in";
         }
-        if (exception instanceof DisabledException || exception instanceof LockedException || exception instanceof BadCredentialsException || exception instanceof CredentialsExpiredException || exception instanceof ApiException) {
+        if(exception instanceof DisabledException || exception instanceof LockedException || exception instanceof BadCredentialsException || exception instanceof CredentialsExpiredException || exception instanceof  ApiException) {
             return exception.getMessage();
         }
-        if (httpStatus.is5xxServerError()) {
-            return "Internal server error occurred";
+        if(httpStatus.is5xxServerError()) {
+            return "An internal server error occurred";
         } else {
-            return "An error occurred occurred. Please try again";
+            return "An error occurred. Please try again.";
         }
     };
 
     private static Response getErrorResponse(HttpServletRequest request, HttpServletResponse response, Exception exception, HttpStatus status) {
         response.setContentType(APPLICATION_JSON_VALUE);
         response.setStatus(status.value());
-        return new Response(now().toString(), status.value(), request.getRequestURI(), HttpStatus.valueOf(status.value()), errorReason.apply(exception, status), getRootCauseMessage(exception), Collections.emptyMap());
+        return new Response(now().toString(), status.value(), request.getRequestURI(), HttpStatus.valueOf(status.value()), errorReason.apply(exception, status), getRootCauseMessage(exception), emptyMap());
     }
 }
